Warnings cleanup for AT drivers

+++ lib/main/AT32F43x/drivers/src/at32f435_437_crm.c
@@ -936,7 +936,7 @@ error_status crm_pll_parameter_calculate(crm_pll_clock_source_type pll_rcs, uint
   }
 
   /* polling pll parameters */
-  for(ms = ms_min; ms <= ms_max; ms ++)
+  for(ms = ms_min; ms <= (int32_t)ms_max; ms ++)
   {
     for(fr = 5; fr >= 0; fr --)
     {
